requirements.txt
requests
python-telegram-bot==21.1.1
